requirements.txt
Django>=6.0,<7.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.6.0
djangorestframework
markdown
django-filter
